refactor: services -> utils
